Splitting `construct`
